requirements.txt
redis==2.10.6
boto3==1.6.16
pygithub
